Move password stuff into its own test case.
